pyproject.toml
[project]
name = "unofficial-reddit-api"
version = "0.1.0"
description = "An unofficial REST API for Reddit that mimics the official Reddit API by scraping data from Reddit."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Anon Coder", email = "anoncoder1337@proton.me" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "hypercorn",
    "playwright",
    "python-dotenv",
    "pydantic"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI"
]

[project.urls]
"Homepage" = "https://codeberg.org/AnonCoder1337/UnofficialRedditAPI"
"Bug Tracker" = "https://codeberg.org/AnonCoder1337/UnofficialRedditAPI/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src_backend", "src_frontend"]

[project.scripts]
unofficial-reddit-api = "src_frontend.main:main"

[project.entry-points.unofficial_reddit_api]
backend = "src_backend.main:main"